Add routing example to the Navbar component
