+++ frontend/src/components/Layout.jsx
@@ -1,0 +1,9 @@
+import { Outlet } from 'react-router-dom'
+import Navbar from './Navbar'
+
+export const Layout = () => (
+    <>
+        <Navbar />
+        <Outlet />
+    </>
+)
