+++ frontend/src/pages/MaterialPage.jsx
@@ -1,0 +1,9 @@
+import { Typography } from '@mui/material'
+
+export const MaterialPage = () => {
+    return (
+        <>
+            <Typography variant="h1">Material Page</Typography>
+        </>
+    )
+}
